requirements.txt
# Core API and web framework
fastapi==0.104.1
uvicorn==0.24.0

# Data processing and analysis
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
pyarrow==14.0.2

# Configuration and validation
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
statsmodels==0.14.0

# Technical analysis (optional - may need system ta-lib)
# ta-lib==0.4.28  # Uncomment if you have ta-lib system library installed
pandas_ta==0.3.14b

# WebSocket and networking
websocket-client==1.6.4
httpx==0.25.2
aiohttp==3.9.1

# Scheduling
APScheduler==3.10.4
pytz==2023.3

# Visualization and dashboard  
streamlit==1.29.0
matplotlib==3.8.2

# Trading and market data
ccxt>=4.0.0
